ai-diet-agent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
supabase==2.12.0
openai==1.51.2
python-dotenv==1.0.1
python-multipart==0.0.9

